Ajustes en Sensor y SensorService

+++ src/main/java/grupo1/caso_practico/service/SensorService.java
@@ -11,6 +11,8 @@ import java.util.List;
 import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
 
+import ch.qos.logback.classic.Logger;
+
 
 @Service
 public class SensorService {
@@ -25,6 +27,7 @@ public class SensorService {
     }
 
     public List<SensorDTO> findAll() {
+        logger.info("Fetching all sensors");
         final List<Sensor> sensors = sensorRepository.findAll(Sort.by("id"));
         return sensors.stream()
                 .map(sensor -> mapToDTO(sensor, new SensorDTO()))
@@ -32,25 +35,39 @@ public class SensorService {
     }
 
     public SensorDTO get(final Long id) {
+        logger.info("Fetching sensor with id {}", id);
         return sensorRepository.findById(id)
                 .map(sensor -> mapToDTO(sensor, new SensorDTO()))
-                .orElseThrow(NotFoundException::new);
+                .orElseThrow(() -> {
+                    logger.error("Sensor with id {} not found", id);    
+                    return new NotFoundException();
+                });
     }
 
     public Long create(final SensorDTO sensorDTO) {
+        logger.info("Creating sensor");
         final Sensor sensor = new Sensor();
         mapToEntity(sensorDTO, sensor);
         return sensorRepository.save(sensor).getId();
     }
 
     public void update(final Long id, final SensorDTO sensorDTO) {
+        logger.info("Updating sensor with id {}", id);
         final Sensor sensor = sensorRepository.findById(id)
-                .orElseThrow(NotFoundException::new);
+            .orElseThrow(() -> {
+                logger.error("Sensor not found with id: {}", id); 
+                return new NotFoundException();
+            });
         mapToEntity(sensorDTO, sensor);
         sensorRepository.save(sensor);
     }
 
     public void delete(final Long id) {
+        logger.info("Deleting sensor with id {}", id);
+        if(sensorRepository.deleteById(id)){
+            logger.error("Sensor not found with id: {}", id); 
+            throw new NotFoundException();
+        }
         sensorRepository.deleteById(id);
     }
